requirements.txt
asgiref==3.5.2
channels==3.0.5
Django==3.2.13
django-crispy-forms==1.14.0
Pillow==9.1.1
pyserial==3.5
python-barcode==0.14.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.2.0